requirements.txt
torch>=1.8.1, <=1.13.0
lightning>=2.0.0rc0
setuptools==67.4.0
seaborn
python-dotenv
ipython[notebook]>=8.0.0, <8.12.0
torchmetrics>=0.7, <0.12
torchvision
pytorch-lightning>=1.4, <2.0.0
matplotlib
google-cloud-storage
tensorboard
google-cloud-aiplatform[tensorboard]
torchserve
torch-model-archiver
typing-extensions
python-json-logger